requirements.txt
langchain
langchain-community
langchain-google-genai
faiss-cpu
streamlit
python-dotenv
pypdf
tiktoken
google-generativeai
os